Finally reworked autocomplete view
flds is now a property and includes an integrity check.
get_queryset may return a list instead of a queryset. 
The list consists of three seperate lists: 1) exact, 2) startswith, 3) contains queryset results.
Due to not returning a queryset object, method has_add_permission has had to be changed as well.

Removed ACList view since it was not used anymore.