preserved_filters changes
A change form's initial data depends heavily on the _changelist_filters that are, by default, preserved from the last 'instance' of the changelist the user visited.
If the user makes changes to the data provided by initial, these changes would be overriden again by the preserved filters. In order to... actually preserve the initial data rather than the _changelist_filters, we *change* the changelist filters according to the last valid form saved.