Registry rework
A HelpRegistry instance provides urls for all 'helptexts' registered to it (get_urls like an AdminSite). This removes the need to update urls.py for every new help page; instead register a helptext and the help page is automatically created.
It also removes the need to create custom views for different FormView help pages.