Changed how audio constructs release_id/discogs_url
The model now only validates that a given url is from discogs domain. 
The form is now responsible for checking if release_id and the id in a discogs_url are the same. The form also constructs a release_id from discogs_url if release_id was missing and vice versa.