Allows to select two kinds of fields on the DuplicateFieldsSelectForm: - one to find the duplicates with - one to display in the found duplicates list
Note that this increases the amount of queries A LOT: each duplicate instance can have different values in the to-be-displayed fields and thus one query for every instance is required whereas before all instances of a duplicates group shared the same duplicate values (those values being the same is what makes them duplicates). Need to reconsider?