test_actions fixed test_permissions_required methods
Trying to access an action that the user does not have access to does not result in 403 - but in an admin message 'No action selected'. Previously, the 'desired' 403 error occurred, not because the user could not access the action, but because they could not access the changelist (no view or change permissions).