Set dbentry app default auto field
The global setting DEFAULT_AUTO_FIELD was set to 'AutoField', which messed with installed third-party apps that used a different default for the auto field (such as 'BigAutoField') by trying to migrate their tables to the global default auto field. By setting the default in the app only, no such migrations will be created.